Fix login command.
In b7ef2ba9d I tried to make the script show help if no command was passed.
However, for some reason this also made it show help if the login command
was run, presumably because login takes no arguments.

This commit uses a different approach to show the help message if there is
no command passed, and makes login work again.